Add hasSeparateLockStore method to RedisStore

+++ src/Illuminate/Cache/RedisStore.php
@@ -11,6 +11,7 @@ use Illuminate\Redis\Connections\PredisClusterConnection;
 use Illuminate\Redis\Connections\PredisConnection;
 use Illuminate\Support\LazyCollection;
 use Illuminate\Support\Str;
+use RuntimeException;
 
 class RedisStore extends TaggableStore implements FlushableLock, LockProvider
 {
@@ -294,9 +295,15 @@ class RedisStore extends TaggableStore implements FlushableLock, LockProvider
      * Remove all locks from the store.
      *
      * @return bool
+     *
+     * @throws \RuntimeException
      */
     public function flushLocks(): bool
     {
+        if (! $this->hasSeparateLockStore()) {
+            throw new RuntimeException('Flushing locks is only supported when the lock store is separate from the cache store.');
+        }
+
         $this->lockConnection()->flushdb();
 
         return true;
@@ -543,5 +550,15 @@ class RedisStore extends TaggableStore implements FlushableLock, LockProvider
         }
 
         return $this->unserialize($value);
+    }
+
+    /**
+     * Determine if the lock store is separate from the cache store.
+     *
+     * @return bool
+     */
+    public function hasSeparateLockStore(): bool
+    {
+        return $this->lockConnection !== $this->connection;
     }
 }
